pyproject.toml
[tool.poetry]
name = "redbeat-example"
version = "0.1.0"
description = ""
authors = ["Hung Hoang <hxuanhung@gmail.com>"]
readme = "README.md"
packages = [{include = "redbeat_example"}]

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.0"
celery = "^5.3.1"
flower = "^2.0.1"
requests = "^2.31.0"
celery-redbeat = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
